% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rpact.as251}
\alias{rpact.as251}
\alias{rpact.as251-package}
\title{rpact AS 251.1}
\description{
TODO
}
\details{
TODO
}
\references{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.rpact.org}
  \item \url{https://www.rpact.com}
  \item \url{https://github.com/rpact-com/rpact.as251}
  \item \url{https://rpact-com.github.io/rpact.as251/}
  \item Report bugs at \url{https://github.com/rpact-com/rpact.as251/issues}
}

}
\author{
Till Jensen, Friedrich Pahlke
}
