% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{rpact.as251}
\alias{rpact.as251}
\alias{rpact.as251-package}
\title{rpact AS 251}
\description{
This R package integrates Charles Dunnetts AS 251 FORTRAN 77 algorithm from the journal article 
"Multivariate Normal Probability Integrals with Product Correlation Structure" in R using Rcpp.
}
\details{
TODO
}
\references{
Dunnett, C. W. (1989). Algorithm AS 251: Multivariate Normal Probability Integrals with Product 
Correlation Structure. Journal of the Royal Statistical Society. 
Series C (Applied Statistics), 38(3), 564–579. 
[doi.org/10.2307/2347754](https://doi.org/10.2307/2347754)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.rpact.org}
  \item \url{https://www.rpact.com}
  \item \url{https://github.com/rpact-com/rpact.as251}
  \item \url{https://rpact-com.github.io/rpact.as251/}
  \item Report bugs at \url{https://github.com/rpact-com/rpact.as251/issues}
}

}
\author{
Till Jensen, Friedrich Pahlke
}
